Adds sorting to observations and renames onservations files
