+++ www/frontends/compiler_gym/src/components/StateVisualizations/InstcountsStateContainer.js
@@ -8,26 +8,24 @@ import React from "react";
 import classnames from "classnames";
 import Highcharts from "highcharts";
 import HighchartsReact from "highcharts-react-official";
-import HC_exporting from "highcharts/modules/exporting";
-import HC_exportData from "highcharts/modules/export-data";
 
-HC_exporting(Highcharts);
-HC_exportData(Highcharts);
+const InstcountsStateContainer = ({ instcount, darkTheme }) => {
 
-const AutophaseStateContainer = ({ autophase, darkTheme }) => {
-
-  const chartData = Object.entries(autophase).map(([category, result]) => ({
-    category,
-    result,
-  }));
+  const chartData = Object.entries(instcount)
+    .map(([category, result]) => ({
+      category,
+      result,
+    }))
+    .sort((a, b) => (a.result > b.result ? -1 : 1));
 
   const options = {
+    colors: ["#64E572", "#DDDF00", "#FF9655", "#FFF263", "#6AF9C4"],
     chart: {
       type: "bar",
       height: 150 + "%", //3:2 ratio
     },
     title: {
-      text: "AutoPhase",
+      text: "InstCounts",
       style: {
         color: darkTheme && "white",
       },
@@ -63,17 +61,20 @@ const AutophaseStateContainer = ({ autophase, darkTheme }) => {
         dataLabels: {
           enabled: true,
           format: "{point.y}",
+          borderWidth: 0,
           color: darkTheme && "white",
+          style: {
+            fontFamily: "menlo",
+          },
         },
       },
     },
     credits: {
       enabled: false,
     },
-
     series: [
       {
-        name: "Autophase",
+        name: "InstCounts",
         data: chartData.map((i) => i.result),
       },
     ],
@@ -99,4 +100,4 @@ const AutophaseStateContainer = ({ autophase, darkTheme }) => {
   );
 };
 
-export default AutophaseStateContainer;
+export default InstcountsStateContainer;
