+++ www/frontends/compiler_gym/src/components/StateVisualizations/index.js
